Reorganize function implementations
* rename `Function::new` to `NativeFunction::new`
* rename `Function::new_self_call` to `WrappedMethod::new`
* rename `Function::new_def` to `Def::new`
* move `Def` and `eval_def` to `crate::eval::def`
